% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean App Data}
\usage{
clean_data(.data)
}
\arguments{
\item{.data}{A data frame. Should be data provided by [get_data()].}
}
\value{
A data frame.
}
\description{
Cleans app data for use in the shiny application.
}
