% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColoradoAwardedDegrees-package.R
\docType{package}
\name{ColoradoAwardedDegrees-package}
\alias{ColoradoAwardedDegrees}
\alias{ColoradoAwardedDegrees-package}
\title{ColoradoAwardedDegrees: Colorado Awarded Degrees}
\description{
A Shiny application that explores degrees awarded in Colorado by higher education institutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSUDenver-Stats-Science-Club/ColoradoAwardedDegrees.git}
  \item Report bugs at \url{https://github.com/MSUDenver-Stats-Science-Club/ColoradoAwardedDegrees.git/issues}
}

}
\author{
\strong{Maintainer}: Brady Lamson \email{blamson@msudenver.edu}

}
\keyword{internal}
