% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dashboard_page.R
\name{dashboard_page}
\alias{dashboard_page}
\title{Dashboard Page for CAD}
\usage{
dashboard_page()
}
\value{
A list.
}
\description{
Contains all dashboard page and body elements used within the UI of the dashboard.
}
