% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_data}
\alias{get_data}
\title{Get App Data}
\usage{
get_data()
}
\value{
A data frame. Contains app data.
}
\description{
A function to send requests to the Colorado Information Marketplace. This will access app data for use in the Shiny dashboard.
}
\references{
For more information on our data visit: https://data.colorado.gov/resource/hxf8-ab6k
}
